MB-69328: Wait for snapshot to be ready after file based backfill
We must wait for the snapshot to be in a state that memcached consider
ready for rebalance to continue. This means that caches will be
transferred or potentially warmed-up such that a takeover is ready to
be performed.

Currently only cache transfer may happen, but the stat has been named
generically such that alternatives (i.e. warmup) could take the place
of it.

Cache transfer relies on the stream that we implicitly create at step
3 of FBR (see comment in ns_single_vbucket_mover.erl).

Change-Id: I230c363ec7b64a06b17660142f86d66b01dc5e2c

Reviewed-on: https://review.couchbase.org/c/ns_server/+/236736
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>