MB-69103: Don't send dup terse cluster info
With this change we no longer will send terse cluster info to memcached
if the info hasn't changed.

This will happen if there are multiple bucket_info_cache_invalidations
and the timing of them was such that a "refresh" might be received by
the handler while a prior one was being handled and nothing in the
"content" had changed.

Change-Id: If40d7f1df0b90d4600cc6fb3cc18d54c57c88b8f
Reviewed-on: https://review.couchbase.org/c/ns_server/+/236752
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>