MB-67395: Add azure domains to encryption settings
The allowed azure domain list is unlikely to change, but it is
best to have it configurable just in case a new domain needs to be
allowed in the future.

This change adds it under /settings/security. Removals, although
unlikely to be needed, are allowed from the list of allowed domains
only if an existing azure key does not exist that uses the domain.

Change-Id: Ia9a2312578c1651f8b731f0db6db79af08a98237
Reviewed-on: https://review.couchbase.org/c/ns_server/+/236817
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>