MB-69348: Upgrade alerts config correctly from now on
For the vast majority of cluster-wide config keys, we correctly
initialise them to the value reflecting MIN_SUPPORTED_VERSION, rather
than LATEST_VERSION_NUM. This ensures that when a new node is added to
the cluster, each unchanged config key is either the same as the
existing default value, or the existing cluster has already upgraded
that key at least once before, ensuring that the existing value is kept
for the mixed version period of the cluster. Once the cluster is fully
upgraded, and the cluster compat mode changes, the upgrade to any
cluster-wide config keys gets performed, ensuring consistent behaviour.

This is quite a subtle technicality, so it's easiest to see what the
above means through seeing an example where it goes wrong.

The email_alerts config key has been using the latest list of alerts
for the 'alerts' and 'pop_up_alerts' sub-keys, rather than reflecting
the MIN_SUPPORTED_VERSION.

- This means that when adding a new node to a cluster when these keys
  have changed, if the existing config key hasn't been
  upgraded/modified, then neither the new or old will have vclocks, so
  conflict resolution will be performed by value instead

- This conflict resolution is arbitrary, which means that in practice,
  the new value was not winning when a new alert was added on the end,
  and so it appeared to behave correctly. Then, when an upgrade removed
  an alert from the list, it caused that new default to be the winning
  value, causing the upgrade to be performed prior to cluster compat
  mode change

- Once the config key has been upgraded by a cluster compat change, or
  manually modified by a user, it will get a vclock added, and will
  from then on always win the conflict resolution against a new default
  value

If we had instead used MIN_SUPPORTED_VERSION, then we would not have
this issue, because the new config would always either be the same as
the existing cluster's config, or the existing cluster's config would
have already been upgraded, and so have a vclock to protect it from
conflict with the new default. We cannot have a not-yet-upgraded config
in the existing cluster at the same time that the default config has
changed, because in that case the MIN_SUPPORTED_VERSION would have been
increased, so in order for the new node to be able to join the existing
cluster, it must have already been upgraded at least to that new
MIN_SUPPORTED_VERSION, which would add the vclock.

So, with all of the above understood, it's clear that we need to start
treating the current email_alerts default config, as of the last
release that it was updated in (7.9), as a fixed value, which should
not be updated again until MIN_SUPPORTED_VERSION >= 8.0.
Since TOTORO hasn't yet been released, we can move the newly added
alerts out of the default config, keeping them only in the upgrade
paths, until MIN_SUPPORTED_VERSION >= 8.1.

In order to make this change, we have to split the alert_keys()
functions out into an alert_keys_default() and alert_keys_all(), so
that we can still get the list of all alerts known at the current
version, which used to just be the same as the default list (plus any
normally disabled-by-default alerts).

This commit also updates the ns_config_default:default/1 comment to
try to make the expected process for updating it clearer, and
explicitly call out that cluster-wide default config should reflect
MIN_SUPPORTED_VERSION.

Change-Id: I77f12efef9c10eec38f44e8f46a4772706e78dd9
Reviewed-on: https://review.couchbase.org/c/ns_server/+/236917
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>