MB-69650: Fix wrong auth info version notifys
In https://review.couchbase.org/c/ns_server/+/209891 the check that
made sure the version of the info being notified matched the worker was
lost.

This leads to crashes when attempting to notify with the wrong version
auth info.

Adds back the version check, and adds a unit test to explicitly cover
this issue.

Change-Id: I4a8700d1c58275785a65008d9dcb4459a99e421e
Reviewed-on: https://review.couchbase.org/c/ns_server/+/237268
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>