MB-69538: Automated tests for uploaders assignment
Test failure is triggered if cluster is not downsizing and
some uploaders are scheduled to upload from scratch. The reason
for it is that uploading from scratch on downsizing clusters
is sometimes unavoidable.

The trivial example is [a, b, c] -> [a, b, undefined], where c
is an uploader. But there could be others.

Test failure is not triggered if the uploaders map is unbalanced.
This is because sometimes it just has to happen. For example it
might happen when you just add nodes to the cluster,
the existing nodes get less uploaders.

The report for all performed tests is generated at the end.
This doesn't affect regression testing, but I need it to explore
various scenarios during the manual testing.

Change-Id: I3b62f82af35536f7851a687dc60ac8351d94e0e5
Reviewed-on: https://review.couchbase.org/c/ns_server/+/237751
Tested-by: Artem Stemkovski <artem@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>