MB-69758: Make rebalance tests mock vbucket movement
Such that we can test it in the future.

This is accomplished by removing the desired_servers from the bucket
config creation helper. Adding this was likely a mistake, it invokes
the bucket placer code which is not normally run, and causes the
rebalance to behave differently (i.e. the bucket is not re-placed on
new nodes without other changes). This brings the code more in line
with typical deployments.

Change-Id: Ic59f5ba7293b5c6a970c10ab935a51d8f371f51d
Reviewed-on: https://review.couchbase.org/c/ns_server/+/237711
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>