MB-69656: Fix decrypt post key rotation in Azure
Azure key vault expects the client to track the key version with the
encrypted data, and use it during decrypt. If this is not done, decrypt
will fail after key gets rotated, and we are trying to decrypt data
that was decrypted with a previous version.

Change-Id: If8c9b9cc4155502d9f71fd39b5c684bad6711eac
Reviewed-on: https://review.couchbase.org/c/ns_server/+/237741
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>