MB-69612: Fix client cert auth during rebalance
We will only specify the client cert info in the download
request to memcached, when client cert auth state is hybrid
or mandatory. When the cert info is not passed, it indicates
to memcached that no client cert auth should be done on
connection memcached initiates for memcached to mecached
file download.

Change-Id: I07ba916a5076e98ec07344cd75816b9474c37c25
Reviewed-on: https://review.couchbase.org/c/ns_server/+/237800
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>